+++ core/lib/moe_core_controller.dart
@@ -1,21 +1,19 @@
 import 'package:flutter/services.dart';
-import 'package:moengage_flutter/in_app_payload_mapper.dart';
-import 'package:moengage_flutter/model/permission_result.dart';
-import 'package:moengage_flutter/moe_android_core.dart';
-import 'package:moengage_flutter/moe_ios_core.dart';
 import 'package:moengage_flutter/constants.dart';
 import 'package:moengage_flutter/core_instance_provider.dart';
+import 'package:moengage_flutter/in_app_payload_mapper.dart';
 import 'package:moengage_flutter/model/inapp/click_data.dart';
 import 'package:moengage_flutter/model/inapp/inapp_data.dart';
 import 'package:moengage_flutter/model/inapp/self_handled_data.dart';
+import 'package:moengage_flutter/model/permission_result.dart';
 import 'package:moengage_flutter/model/push/push_campaign_data.dart';
 import 'package:moengage_flutter/model/push/push_token_data.dart';
+import 'package:moengage_flutter/moe_android_core.dart';
 import 'package:moengage_flutter/moe_cache.dart';
+import 'package:moengage_flutter/moe_ios_core.dart';
 import 'package:moengage_flutter/moengage_flutter.dart';
 import 'package:moengage_flutter/push_payload_mapper.dart';
 import 'package:moengage_flutter/utils.dart';
-
-import 'constants.dart';
 
 class CoreController {
   String _tag = "${TAG}CoreController";
@@ -34,7 +32,7 @@ class CoreController {
   factory CoreController() => _instance;
 
   Future<dynamic> _handler(MethodCall call) async {
-    print("$_tag _handler() : Received callback. Payload " + call.method);
+    Logger.v("$_tag _handler() : Received callback. Payload ${call.method}");
     try {
       if (call.method == callbackPushTokenGenerated) {
         PushTokenData? data =
@@ -97,12 +95,12 @@ class CoreController {
       if (call.method == callbackOnInAppSelfHandled) {
         SelfHandledCampaignData? data =
             InAppPayloadMapper().selfHandledCampaignFromJson(call.arguments);
-        print("$_tag _handler() : data: $data");
+        Logger.i("$_tag _handler() : data: $data");
         if (data != null) {
           SelfHandledInAppCallbackHandler? handler = CoreInstanceProvider()
               .getCallbackCacheForInstance(data.accountMeta.appId)
               .selfHandledInAppCallbackHandler;
-          print("$_tag _handler() : handler: $handler");
+          Logger.v("$_tag _handler() : handler: $handler");
           if (handler != null) {
             handler.call(data);
           }
@@ -116,8 +114,9 @@ class CoreController {
           handler.call(data);
         }
       }
-    } catch (e) {
-      print("$_tag Error: ${call.toString()} has an Exception: $e");
+    } catch (e, stackTrace) {
+      Logger.e("$_tag Error: ${call.toString()} has an Exception:",
+          error: e, stackTrace: stackTrace);
     }
   }
 }