+++ inbox/lib/payload_transformer.dart
@@ -1,9 +1,9 @@
 import 'dart:convert';
-import 'dart:ffi';
-import 'dart:io';
 
+import 'package:moengage_flutter/internal/logger.dart';
 import 'package:moengage_inbox/action.dart';
 import 'package:moengage_inbox/action_type.dart';
+import 'package:moengage_inbox/constants.dart';
 import 'package:moengage_inbox/inbox_data.dart';
 import 'package:moengage_inbox/inbox_message.dart';
 import 'package:moengage_inbox/media.dart';
@@ -11,7 +11,8 @@ import 'package:moengage_inbox/media_type.dart';
 import 'package:moengage_inbox/navigation_action.dart';
 import 'package:moengage_inbox/navigation_type.dart';
 import 'package:moengage_inbox/text_content.dart';
-import 'package:moengage_inbox/constants.dart';
+
+final _tag = "${TAG}PayloadTransformer";
 
 // Unclicked Count
 int fetchUnclickedCount(dynamic unClickedPayload) {
@@ -51,8 +52,8 @@ InboxData? deSerializeInboxMessages(dynamic messagesPayload) {
     Map<String, dynamic> dataPayload = message[keyData];
     return InboxData(
         dataPayload[keyPlatform], messagesJsonToList(dataPayload[keyMessages]));
-  } catch (e) {
-    print(e);
+  } catch (e, stackTrace) {
+    Logger.e("$_tag Error: ", error: e, stackTrace: stackTrace);
   }
   return null;
 }
@@ -83,8 +84,9 @@ InboxMessage? messageFromJson(Map<String, dynamic> message) {
         message[keyReceivedTime],
         message[keyExpiryTime],
         message[keyPayload]);
-  } catch (e) {
-    print(e);
+  } catch (e, stacktrace) {
+    Logger.e("$_tag Error: messageFromJson InboxMessage ",
+        stackTrace: stacktrace);
   }
   return null;
 }
@@ -126,8 +128,9 @@ Action? actionFromMap(Map<String, dynamic> actionMap) {
   switch (actionType) {
     case ActionType.navigation:
       return navigationActionFromMap(actionType, actionMap);
+    default:
+      return null;
   }
-  return null;
 }
 
 NavigationAction navigationActionFromMap(
@@ -167,8 +170,9 @@ Map<String, dynamic>? actionToMap(Action action) {
   switch (action.actionType) {
     case ActionType.navigation:
       return navigationActionToMap(action as NavigationAction);
+    default:
+      return null;
   }
-  return null;
 }
 
 Map<String, dynamic> navigationActionToMap(NavigationAction navigationAction) {