+++ core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # MoEngage Flutter Plugin
 
+# 08-02-2023
+## 5.3.0
+- Security improvement: controlled logging for release, debug and profile mode
+- Android
+    - Support for 2 Step Push Optin InApps
+    - Device Id enable / disable support
+- iOS
+    - MoEngage-iOS-SDK version updated to `~>9.4.0`.
+
 # 23-01-2023
 ## 5.2.0
 - iOS