+++ inbox/CHANGELOG.md
@@ -1,5 +1,11 @@
 # MoEngage Flutter Inbox Plugin
 
+# 08-02-2023
+## 4.3.0
+- Security improvement: controlled logging for release, debug and profile mode
+- iOS
+    - MoEngageInbox SDK version updated to `~>2.4.0`.
+
 # 23-01-2023
 ## 4.2.0
 - iOS